Keep clearing the queue at its end
Rewinding a finished queue reads badly in the UI: the queue panel resolves the
current row with `current_index ?? 0`, so a null index highlights the first track
as "Now playing" with a divider above it, and the bottom bar falls back to the
player name because the "queue empty" caption is gated on an item count of zero.

Clearing keeps the finished state unambiguous: the queue reports no items, the
existing empty-queue caption renders, and the source advertises play as
unavailable. Stating "this queue has ended" more explicitly needs a queue-level
flag plus frontend support, which is follow-up work.
