+++ music_assistant/controllers/player_queues/base.py
@@ -53,7 +53,6 @@ class _PlayerQueuesBase(CoreController):
         def index_by_id(self, queue_id: str, queue_item_id: str) -> int | None: ...
         def get_next_item(self, queue_id: str, cur_index: int | str) -> QueueItem | None: ...
         def clear(self, queue_id: str, skip_stop: bool = False) -> None: ...
-        def reset_playback_position(self, queue_id: str) -> None: ...
         def signal_update(self, queue_id: str, items_changed: bool = False) -> None: ...
         def on_player_update(
             self, player: Player, changed_values: dict[str, tuple[Any, Any]]
